% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_anno.R
\name{AnnoKn}
\alias{AnnoKn}
\title{Annotation-informed Knockoffs}
\usage{
AnnoKn(X, Xk, y, lams = NULL, attempts = NULL, R)
}
\arguments{
\item{X}{Orignal design matrix}

\item{Xk}{Knockoff copies of X}

\item{y}{response vector}

\item{R}{matrix of standardized annotation matrix}
}
\value{
A list contains the following results: Prior_final, final weight phi_j; beta, final beta in adaptive lasso; lambda_s, estimated lambda_l parameter;
}
\description{
Annotation-informed Knockoffs
}
