% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_anno.R
\name{AnnoKn_lite}
\alias{AnnoKn_lite}
\title{Annotation-informed Knockoffs}
\usage{
AnnoKn_lite(X, Xk, y, R)
}
\arguments{
\item{X}{Orignal design matrix}

\item{Xk}{Knockoff copies of X}

\item{y}{response vector}

\item{R}{matrix of standardized annotation matrix}
}
\value{
A list contains the following results: Prior, weight phi_j; beta, beta in adaptive lasso; lambda_s, estimated lambda_l parameter;
}
\description{
Annotation-informed Knockoffs
}
