% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_anno.R
\name{fdr_cal}
\alias{fdr_cal}
\title{Calculate FDR}
\usage{
fdr_cal(index_result, index_true)
}
\arguments{
\item{index_result}{index for selected covariates}

\item{index_true}{index for true relevant covariates}
}
\value{
realized FDR
}
\description{
Calculate FDR
}
