% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.filter}
\alias{GhostKnockoff.filter}
\title{GhostKnockoff Filtering}
\usage{
GhostKnockoff.filter(T_0, T_k, clusters = 1:length(T_0))
}
\arguments{
\item{T_0}{Vector or Matrix, feature importance statistics for the original variables.}

\item{T_k}{Matrix of size (p x M), feature importance statistics for the knockoff copies.}

\item{clusters}{Vector of size p, indicating cluster assignments for features.
Defaults to 1:length(T_0).}
}
\value{
A list containing:
\itemize{
\item \code{kappa}: Vector of signs indicating whether the original variable (0) or
which knockoff copy (1 to M) performed best.
\item \code{tau}: Vector of differences between the max statistic and the
median of the other statistics.
\item \code{q}: Vector of estimated q-values for FDR control.
}
}
\description{
Computes knockoff statistics (\eqn{\kappa} and \eqn{\tau}) and performs
FDR control to identify significant features.
}
