% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GhostKnockoff.R
\name{GhostKnockoff.fit}
\alias{GhostKnockoff.fit}
\title{GhostKnockoff Fitting}
\usage{
GhostKnockoff.fit(
  Zscore_0,
  N.effect,
  fit.prelim,
  method = "susie",
  type = "fdr",
  M.fwer = 50
)
}
\arguments{
\item{Zscore_0}{Vector or Matrix, observed Z-scores from the original study.}

\item{N.effect}{Vector, sample sizes of the original study.}

\item{fit.prelim}{List, the output from GhostKnockoff.prelim.}

\item{method}{String, selection method for feature importance scores. Options: 'marginal', 'lasso', or 'lasso.approx.lambda'.}

\item{type}{String, choice of FDR control type. Options: 'fdr' or 'fwer'.}

\item{M.fwer}{Integer, number of copies used if type is 'fwer'. Default is 50.}
}
\value{
A list containing:
\itemize{
\item \code{T_0}: Feature importance statistics for original variables.
\item \code{T_k}: Matrix of statistics for knockoff copies.
\item \code{kappa}: Calculated kappa statistic.
\item \code{tau}: Calculated tau statistic.
\item \code{temp.A}: Adjusted covariance matrix used in fitting.
\item \code{r_all}: Concatenated vector of statistics.
\item \code{lambda.seq}: Lambda sequence used in the Lasso path.
\item \code{lambda}: The finally selected lambda value.
}
}
\description{
GhostKnockoff Fitting
}
