% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoGK.R
\name{GK_anno_M}
\alias{GK_anno_M}
\title{Annotation-informed GhostKnockoff}
\usage{
GK_anno_M(Z, R, M = 1, LD, n, ts = "lasso")
}
\arguments{
\item{Z}{vector of z-scores}

\item{R}{matrix of standardized annotation matrix}

\item{M}{number of knockoff copies}

\item{LD}{correlation matrix}

\item{n}{sample size}
}
\value{
A list containing the following results:
\itemize{
\item \code{T_0} and \code{T_k}: Feature statistics for original data and knockoff copies, same as GhostKnockoff.
\item \code{CV_BEST}: Best cross-validation score.
\item \code{lambda_BEST}: The optimal lambda parameter chosen.
\item \code{lambda_s}: Estimated lambda_s parameter.
\item \code{Prior}: The calculated prior weights.
}
}
\description{
Annotation-informed GhostKnockoff
}
