% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoGK.R
\name{GK_simple}
\alias{GK_simple}
\title{Simplfied Annotation-informed GhostKnockoff}
\usage{
GK_simple(Z, R, M = 1, LD, n, ts = "lasso")
}
\arguments{
\item{Z}{vector of z-scores}

\item{R}{matrix of standardized annotation matrix}

\item{M}{number of knockoff copies}

\item{LD}{correlation matrix}

\item{n}{sample size}
}
\value{
A list contains results
}
\description{
Simplfied Annotation-informed GhostKnockoff
}
