% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_anno.R
\name{power_cal}
\alias{power_cal}
\title{Calculate power}
\usage{
power_cal(index_result, index_true)
}
\arguments{
\item{index_result}{index for selected covariates}

\item{index_true}{index for true relevant covariates}
}
\value{
statistical power
}
\description{
Calculate power
}
